moved EqSearch.java to a searches package, and adjusted other files accordingly
