+++ src/main/java/hello/controllers/EarthquakesController.java
@@ -12,7 +12,7 @@ import java.util.Map;
 import java.util.HashMap;
 import hello.geojson.FeatureCollection;
 import hello.services.EarthquakeQueryService;
-import hello.EqSearch;
+import hello.searches.EqSearch;
 
 import com.nimbusds.oauth2.sdk.client.ClientReadRequest;
